Move ProductController
