+++ app/Http/Controllers/common/ProductController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\common;
 
 use App\Models\Product;
 use App\Models\ErrorLog;
