+++ routes/api.php
@@ -1,6 +1,6 @@
 <?php
 
-use App\Http\Controllers\ProductController;
+use App\Http\Controllers\common\ProductController;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
